Write `HeapKind` and `BlockType` as signed values (#32)
The heap and block type encodings both use signed values, where negative
values represent well-known values, and non-negative values are indexes
into the type section. This commit fixes a bug where the indexes were
written with LEB128 encoding (instead of SLEB128 encoding).

The difference is first noticeable at index 64, where the unsigned
encoding is `0x40`, but the unsigned encoding is `0xc0 0x00`.

+++ include/wasp/binary/write.h
@@ -104,7 +104,7 @@ Iterator Write(HeapType value, Iterator out) {
     return Write(encoding::HeapKind::Encode(value.heap_kind()), out);
   } else {
     assert(value.is_index());
-    return Write(value.index(), out);
+    return Write(static_cast<s32>(value.index()), out);
   }
 }
 
@@ -153,7 +153,7 @@ Iterator Write(BlockType value, Iterator out) {
     return Write(encoding::BlockType::Void, out);
   } else {
     assert(value.is_index());
-    return Write(value.index(), out);
+    return Write(static_cast<s32>(value.index()), out);
   }
 }
